+++ galois/_codes/_bch.py
@@ -7,9 +7,9 @@ from numba import int64
 import numpy as np
 
 from .. import _lfsr
-from .._fields import Field, FieldArrayClass, FieldArray, GF2, matlab_primitive_poly
+from .._fields import Field, FieldArrayClass, FieldArray, GF2
 from .._overrides import set_module
-from .._polys import Poly
+from .._polys import Poly, matlab_primitive_poly
 from .._prime import factors
 
 from ._cyclic import poly_to_generator_matrix, roots_to_parity_check_matrix
