+++ galois/_polys/__init__.py
@@ -1,5 +1,8 @@
 """
 A subpackage containing arrays over Galois fields.
 """
+from ._functions import *
+from ._irreducible import *
 from ._poly import *
-from ._functions import *  # pylint: disable=redefined-builtin
+from ._primitive import *
+from ._primitive_element import *
