+++ galois/_polys/_conversions.py
@@ -1,6 +1,5 @@
 """
 A module that contains various functions to convert between polynomial strings, coefficients, and integer representations.
-These functions are shared between the _field and _poly subpackages.
 """
 from typing import Tuple, List, Optional
